Update unit representation & tactical actions (#3145)
The goal of this commit is to change unitholders so that damaged units are represented as
- dread=[0,1] (0 undamaged, 1 damaged) instead of
- dread=1, damaged_dread=1 (1 total, 1 of which is damaged)

I updated also several other things to support this, and updated several things because they annoyed me
- Tactical Action buttons were redone
- - Tactical action tracking was also redone on the backend. This introduces a small regression for in progress actions
- Assign Hits buttons were redone
- Repair Damage buttons were redone
- Added "DestroyUnitService"
- - Redid -most- of the spots in the code where we are supposed to be destroying units, and all the destroy handling was moved to this common place to make it easy to use and extend in the future
- - Added some spots in the code where it should be destroy but wasn't
- Added "MoveUnitService"
- Added "CaptureUnitService"

And some misc fixes:
- Fix drumroll not cleaning up messages
- close threads end of round
- cache milty draft tiles
- cache unit keys
- misc code cleanup
- remove Joiner dependency

+++ src/main/java/ti4/buttons/handlers/leader/agent/XxchaAgentButtonHandler.java
@@ -16,10 +16,12 @@ class XxchaAgentButtonHandler {
     @ButtonHandler("xxchaAgentRemoveInfantry_")
     public static void resolveXxchaAgentInfantryRemoval(Player player, Game game, ButtonInteractionEvent event, String buttonID) {
         Player p2 = game.getPlayerFromColorOrFaction(buttonID.split("_")[1]);
+        if (p2 == null) return;
         String planet = buttonID.split("_")[2];
         String planetRep = Helper.getPlanetRepresentation(planet, game);
         ButtonHelper.deleteMessage(event);
         RemoveUnitService.removeUnits(event, game.getTileFromPlanet(planet), game, p2.getColor(), "1 infantry " + planet);
+        ButtonHelper.resolveInfantryRemoval(p2, 1);
         MessageHelper.sendMessageToChannel(player.getCorrectChannel(), player.getRepresentationUnfogged() + " you removed 1 infantry from " + planetRep);
         MessageHelper.sendMessageToChannel(p2.getCorrectChannel(), p2.getRepresentationUnfogged() + " 1 infantry of yours on " + planetRep + " was removed via the Ggrocuto Rinn, the Xxcha agent.");
     }
